LingoLearnAI/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
python-dotenv==1.0.0
requests==2.31.0
googletrans==4.0.0-rc1
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.0
pandas==2.1.1
pytest==7.4.0
gunicorn==21.2.0 